Handle directly upload files by Athena
